Properly tokenize paragraphs with inline code
Add special case for inline code that consumes all enclosed text and ignores sequences like images or custom blocks that would otherwise end the paragraph and inline code token. This is a bit different than the `icode` and `icode2` groups as they don't have to consider line breaks that would end the paragraph.