Add text token definition
An extension can now define custom tokens inside inline text, by providing a regex, group name and token class. The class is added to the token match group table, an token type enum item is added and the token type is set on the class itself.
The regex is currently added to the paragraph and quote tokenizer regex in a very crude way.