Add extension class hook for when to inject token classes
This hook is only called once for the whole program (compiler module instance) and only for extensions actually added to a compiler instance. If an extension is used by any compiler instance its classes are permanently injected into the class hierarchy. This assumes that all compiler instances are configured the same way while the program is running.